+++ assets/style/variables.scss
@@ -131,7 +131,7 @@ $content-br: 6pt;
 $content-height: 35pt;
 $content-padding: 8pt;
 
-$border: 1px solid #C5C5C5;
+$border: 1px solid $color-border;
 $sidebar-width: var(--sidebar-width);
 $botnav-height: 50pt;
